% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{addVar}
\alias{addVar}
\title{Add variables not automatically created i.e. interactions and polynomials}
\usage{
addVar(mf, name, poly = NA)
}
\arguments{
\item{mf}{data.frame}

\item{name}{character}

\item{poly}{integer}
}
\value{
data.frame
}
\description{
Add variables not automatically created i.e. interactions and polynomials
}
