% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getMF}
\alias{getMF}
\title{Make a variable index from a data set}
\usage{
getMF(df, keep = NA, drop = NA)
}
\arguments{
\item{df}{data.frame}

\item{keep}{character}

\item{drop}{character}
}
\value{
data.frame
}
\description{
Make a variable index from a data set
}
