% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{modelSpec}
\alias{modelSpec}
\title{Recall a model specification}
\usage{
modelSpec(mf, resp, model_names)
}
\arguments{
\item{mf}{data.frame}

\item{resp}{character}

\item{model_names}{character}
}
\value{
formula
}
\description{
Recall a model specification
}
