% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{modelVec}
\alias{modelVec}
\title{Recall a vector of variable names}
\usage{
modelVec(mf, model_names)
}
\arguments{
\item{mf}{data.frame}

\item{model_names}{character}
}
\value{
character
}
\description{
Recall a vector of variable names
}
