% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setMod}
\alias{setMod}
\title{Define a model}
\usage{
setMod(mf, model_name, var_names)
}
\arguments{
\item{mf}{data.frame}

\item{model_name}{character}

\item{var_names}{character}
}
\value{
data.frame
}
\description{
Define a model
}
